snowflake - parallel load from manifest FIXE #20
